Modal refactor
