+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/projects/show/products/index.tsx
@@ -1,17 +1,14 @@
 import * as React from 'react';
-import { Modal } from 'semantic-ui-react';
-import CloseIcon from '@material-ui/icons/Close';
-
-import { translate, InjectedTranslateProps as i18nProps } from 'react-i18next';
 import { compose } from 'recompose';
-import { ProjectAttributes } from '@data/models/project';
-import { ResourceObject } from 'jsonapi-typescript';
-import { PROJECTS_TYPE } from '@data';
+
+import { ProjectResource } from '@data';
+import { withTranslations, i18nProps } from '@lib/i18n';
+import ProductModal from './modal';
 
 import './styles.scss';
 
 interface Params {
-  project: ResourceObject<PROJECTS_TYPE, ProjectAttributes>;
+  project: ProjectResource;
 }
 
 type IProps =
@@ -24,33 +21,13 @@ class Products extends React.Component<IProps> {
 
     const { t, project } = this.props;
 
-    const trigger = (
-      <button className='ui button'>
-        {t('project.products.add')}
-      </button>
-    );
-
     return (
       <div className='product'>
         <h3>{t('project.products.title')}</h3>
         <div className='empty-products flex align-items-center justify-content-center'>
           <span>{t('project.products.empty')}</span>
         </div>
-        <Modal
-          trigger={trigger}
-          className='medium products-modal'
-          closeIcon={<CloseIcon className='close-modal'/>}
-        >
-          <Modal.Header>Select products</Modal.Header>
-          <Modal.Content>
-            Products
-          </Modal.Content>
-          <Modal.Actions>
-            <button className='ui button huge'>
-              Done
-            </button>
-          </Modal.Actions>
-        </Modal>
+        <ProductModal />
       </div>
     );
 
@@ -59,5 +36,5 @@ class Products extends React.Component<IProps> {
 }
 
 export default compose(
-  translate('translations')
+  withTranslations
 )(Products);
