+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/projects/show/products/styles.scss
@@ -7,6 +7,7 @@
   line-height: 1.53;
   text-align: center;
   color: $gray;
+  cursor: pointer;
 }
 
 .ui.modal.products-modal > .actions {
